resolve out of boundary data read and write
An array is created to access two bytes (uint8_t) from one unsigned short.
However, the loop count variable is 4 instead of 2 as it uses incorrect data type (unsinged).
It causes 4 iterations of for loops for byte data copies

+++ src/mcmCompiler/src/base/attribute_def/std_array_unsigned_short_4.cpp
@@ -63,7 +63,7 @@ namespace mv
             for(auto v: vec)
             {
                 Tmp tmp = {v};
-                for(unsigned j = 0; j < sizeof(unsigned); ++j)
+                for(size_t j = 0; j < sizeof(unsigned short); ++j)
                     toReturn[i++] = tmp.bytes[j];
             }
             return toReturn;
